ai_helper_api/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==0.19.0
azure-identity
azure-keyvault-secrets